requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pdfminer.six>=20221105
aiohttp>=3.9.0
aiofiles>=23.0.0
mcp>=1.0.0
PyYAML>=6.0
